# Remove `Input.HasNext()`