This PR removes `Input.HasNext()` method. Instead, it introduces a new invariant: `Input.Next()` must return `nil` when the iteration is done. It is built on top of https://github.com/PlanitarInc/go-sitemap/pull/17. It solves the same problem that https://github.com/PlanitarInc/go-sitemap/pull/17 aims to solve but it simplifies the solution a bit by reducing the input interface.

Additionally, it removes `errMaxCapReached` error type used internally. Now that we explicitly return a carry-over entry, it is not needed.

### Benchmarks

No changes compared to https://github.com/PlanitarInc/go-sitemap/pull/17: same allocations, maybe a bit faster because `HasNext()` is gone.

```
$ go test -run '^$' -bench . -benchmem | tee new2.bench
goos: darwin
goarch: amd64
pkg: github.com/PlanitarInc/go-sitemap
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkWriteAll/1-12                      	 1328336	       945.2 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/10-12                     	  150031	      6808 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/100-12                    	   17624	     62340 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/1000-12                   	    1898	    627547 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/10000-12                  	     192	   6315018 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/100000-12                 	      18	  62742897 ns/op	     160 B/op	       4 allocs/op
BenchmarkWriteAll/1000000-12                	       2	 632302752 ns/op	     736 B/op	      22 allocs/op
BenchmarkWriteUrlset/1-12                   	 1720742	       693.7 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/10-12                  	  185407	      6428 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/100-12                 	   19058	     62912 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/1000-12                	    1857	    648849 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/10000-12               	     189	   6213696 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/100000-12              	      37	  31669197 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/1000000-12             	      34	  31461020 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/1-12                    	 9459012	       126.5 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/10-12                   	 1356752	       877.8 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/100-12                  	  137701	      8257 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/1000-12                 	   14607	     84025 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/10000-12                	    1435	    816697 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/100000-12               	     146	   8216622 ns/op	      32 B/op	       1 allocs/op
BenchmarkSitemapWriter_writeXmlString/short-12         	32062255	        38.36 ns/op	       0 B/op	       0 allocs/op
BenchmarkSitemapWriter_writeXmlString/long-12          	 5126026	       211.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkSitemapWriter_writeXmlTime/utc-12             	 5928306	       203.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkSitemapWriter_writeXmlTime/custom-12          	 5191456	       217.9 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/PlanitarInc/go-sitemap	37.201s
```

Comparison to master:

```
$ benchcmp old.bench new2.bench
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                          old ns/op     new ns/op     delta
BenchmarkWriteAll/1-12                             966           945           -2.15%
BenchmarkWriteAll/10-12                            7404          6808          -8.05%
BenchmarkWriteAll/100-12                           84750         62340         -26.44%
BenchmarkWriteAll/1000-12                          822508        627547        -23.70%
BenchmarkWriteAll/10000-12                         7210228       6315018       -12.42%
BenchmarkWriteAll/100000-12                        73317026      62742897      -14.42%
BenchmarkWriteAll/1000000-12                       695820800     632302752     -9.13%
BenchmarkWriteUrlset/1-12                          909           694           -23.65%
BenchmarkWriteUrlset/10-12                         8215          6428          -21.75%
BenchmarkWriteUrlset/100-12                        76855         62912         -18.14%
BenchmarkWriteUrlset/1000-12                       745129        648849        -12.92%
BenchmarkWriteUrlset/10000-12                      7530459       6213696       -17.49%
BenchmarkWriteUrlset/100000-12                     40530718      31669197      -21.86%
BenchmarkWriteUrlset/1000000-12                    38309880      31461020      -17.88%
BenchmarkWriteIndex/1-12                           167           126           -24.34%
BenchmarkWriteIndex/10-12                          1155          878           -24.00%
BenchmarkWriteIndex/100-12                         10980         8257          -24.80%
BenchmarkWriteIndex/1000-12                        97846         84025         -14.13%
BenchmarkWriteIndex/10000-12                       985195        816697        -17.10%
BenchmarkWriteIndex/100000-12                      10004047      8216622       -17.87%
BenchmarkSitemapWriter_writeXmlString/short-12     45.6          38.4          -15.84%
BenchmarkSitemapWriter_writeXmlString/long-12      259           212           -18.14%
BenchmarkSitemapWriter_writeXmlTime/utc-12         249           204           -18.37%
BenchmarkSitemapWriter_writeXmlTime/custom-12      277           218           -21.39%

benchmark                                          old allocs     new allocs     delta
BenchmarkWriteAll/1-12                             3              3              +0.00%
BenchmarkWriteAll/10-12                            3              3              +0.00%
BenchmarkWriteAll/100-12                           3              3              +0.00%
BenchmarkWriteAll/1000-12                          3              3              +0.00%
BenchmarkWriteAll/10000-12                         3              3              +0.00%
BenchmarkWriteAll/100000-12                        4              4              +0.00%
BenchmarkWriteAll/1000000-12                       22             22             +0.00%
BenchmarkWriteUrlset/1-12                          1              1              +0.00%
BenchmarkWriteUrlset/10-12                         1              1              +0.00%
BenchmarkWriteUrlset/100-12                        1              1              +0.00%
BenchmarkWriteUrlset/1000-12                       1              1              +0.00%
BenchmarkWriteUrlset/10000-12                      1              1              +0.00%
BenchmarkWriteUrlset/100000-12                     1              1              +0.00%
BenchmarkWriteUrlset/1000000-12                    1              1              +0.00%
BenchmarkWriteIndex/1-12                           1              1              +0.00%
BenchmarkWriteIndex/10-12                          1              1              +0.00%
BenchmarkWriteIndex/100-12                         1              1              +0.00%
BenchmarkWriteIndex/1000-12                        1              1              +0.00%
BenchmarkWriteIndex/10000-12                       1              1              +0.00%
BenchmarkWriteIndex/100000-12                      1              1              +0.00%
BenchmarkSitemapWriter_writeXmlString/short-12     0              0              +0.00%
BenchmarkSitemapWriter_writeXmlString/long-12      0              0              +0.00%
BenchmarkSitemapWriter_writeXmlTime/utc-12         0              0              +0.00%
BenchmarkSitemapWriter_writeXmlTime/custom-12      0              0              +0.00%

benchmark                                          old bytes     new bytes     delta
BenchmarkWriteAll/1-12                             128           128           +0.00%
BenchmarkWriteAll/10-12                            128           128           +0.00%
BenchmarkWriteAll/100-12                           128           128           +0.00%
BenchmarkWriteAll/1000-12                          128           128           +0.00%
BenchmarkWriteAll/10000-12                         128           128           +0.00%
BenchmarkWriteAll/100000-12                        160           160           +0.00%
BenchmarkWriteAll/1000000-12                       736           736           +0.00%
BenchmarkWriteUrlset/1-12                          32            32            +0.00%
BenchmarkWriteUrlset/10-12                         32            32            +0.00%
BenchmarkWriteUrlset/100-12                        32            32            +0.00%
BenchmarkWriteUrlset/1000-12                       32            32            +0.00%
BenchmarkWriteUrlset/10000-12                      32            32            +0.00%
BenchmarkWriteUrlset/100000-12                     32            32            +0.00%
BenchmarkWriteUrlset/1000000-12                    32            32            +0.00%
BenchmarkWriteIndex/1-12                           32            32            +0.00%
BenchmarkWriteIndex/10-12                          32            32            +0.00%
BenchmarkWriteIndex/100-12                         32            32            +0.00%
BenchmarkWriteIndex/1000-12                        32            32            +0.00%
BenchmarkWriteIndex/10000-12                       32            32            +0.00%
BenchmarkWriteIndex/100000-12                      32            32            +0.00%
BenchmarkSitemapWriter_writeXmlString/short-12     0             0             +0.00%
BenchmarkSitemapWriter_writeXmlString/long-12      0             0             +0.00%
BenchmarkSitemapWriter_writeXmlTime/utc-12         0             0             +0.00%
BenchmarkSitemapWriter_writeXmlTime/custom-12      0             0             +0.00%
```