@korya 2021-08-05
OK. Released `v1.3.0`, please use it.